fix: add error msg when no entity is found for a (theoretically) valid key
If a urlsafe key is queried for for an entity that does not exist, but the key
is valid for the application (for example: in the event that the entity has
been deleted from the datastore), that query will now return a
NotFoundException. Previously, the query would return None, which would result
in an Internal Server Error, since endpoints is expecting to return a GameForm
object.